requirements.txt
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.20.0
Pillow>=8.0.0
pandas>=1.3.0
numpy>=1.20.0
tqdm>=4.60.0
matplotlib>=3.3.0